refactor : 알림초기화 버튼을 목록컴포넌트로 이동
